Add ability for tacacs+ users to login through mapped users
This supports the ability to login via tacacs authentication, without
needing to add tacacs users to the local password file.  For
environments with many tacacs users, this is a significant improvement.

The method used is to add at least one local tacacs user (e.g., tacacs0)
to the password file, and up to 16 (tacacs0..15).  When authenticating
via tacacs, if the tacacs user is not found in the local password file,
the username matching the server-proviced privilege level is used.

That is, if user tacuser37 is at privilege 14, tacacs14 is looked up. If
found, the home directory, uid, gid, etc. for tacacs14 is used.  If not,
tacacs13 is looked up, and so on down to tacacs0.   This allows sparse
mapping if desired.

Most of the work is done in a new package libtacplus-map1 through it's
interfaces get_user_to_auth() and update_mapuser().  There are also 4
other associated packages to provide NSS methods for tacacs, accounting
via an audisp plugin, and a simple tac interface library so that other
shared libraries can interface to libtac.

Quite a bit of re-factoring was also done, including adding development
package for libtac, so it can be used by other packages as a shared
library.

Global symbols that conflicted with other libraries such as xmalloc
were renamed to tac_xmalloc as well.

A sample file Pam.d.common-example was included to make the
configuration simpler.

The ability to look at a secondary file for configuration, mostly for
the tacacs+ server and secrete were added, so configuration does need
to be duplicated in multiple files.

Fixed tacc.1 man page install.  Fixed unprotected reference to magic()
that was left when SSL was added.  Added ability to configure no use
of openssl.  Changed debian default config to not use openssl, because of
GPL violation mixing with openssl.  Fixed tacc.c and crypt.c to compile
and work without SSL as well as with SSL.

Added debian symbols files.  The tac_magic symbol will show up
as missing if SSL is used.

Fixed pam_tacplus.c fix to use correct path for magic.h include

Fixed some compiler warnings, including changing __FUNCTION__ to
__func__ everywhere.

Removed: Pre-Depends: ${misc:Pre-Depends} from debian/control. jessie
dpkg-gencontrol says it's unknown substitution.