Fix fd leak in tac_connect_single(). show tac_author_read() status on fail
There were a number of places in tac_connect_single() that could
leak an fd on errors.  Unify, also gives a unified debug point.

When tac_author_read() fails, show the status in arep.status
for better debugging.

For short reads that aren't an error, don't use %m in syslog,
it produces misleading debugs.  Use different text for short
reads and for errors.