If we get failure status while authenticating, try remaining servers
When there are multiple tacacs servers configured, keep trying other
servers if we successfully talk to a server, but it returns a failure
during authentication.  This can happen when the server doesn't have
info for a userid, and since servers can have different databases,
we need to try all the possible servers.

Also remove unnecessary tacplus_server_t arg passing in a couple of cases.