Be a bit more paranoid about closing fd's
Ticket: none
Reviewed By: nobody
Testing Done: ran normal cases

Since we close tac_fd in various places, need to set it to -1 after,
so we don't close, leave it set, and then close again after other
code (ours or outside pam) opens and gets the same fd.

I know of no cases where this has happened, but as a shared lib, we
need to be paranoid.

Also, avoid a pointless close(-1) in tac_connect_single().  Not a case
where we can incorrectly close a valid fd.